fix: the metadata should not modify in lint
