+++ pkg/api/metadata.go
@@ -1,4 +1,4 @@
-package extension
+package api
 
 import (
 	"encoding/base64"
@@ -6,33 +6,19 @@ import (
 	"mime"
 	"net/http"
 	"os"
-	"path"
+	ospath "path"
 	"strings"
 
 	"github.com/go-playground/validator/v10"
 	"helm.sh/helm/v3/pkg/chart"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/json"
 	corev1alpha1 "kubesphere.io/api/core/v1alpha1"
-	runtimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/yaml"
 
 	"github.com/kubesphere/ksbuilder/pkg/iso639"
 )
 
-var Categories = []string{
-	"ai-machine-learning",
-	"computing",
-	"database",
-	"dev-tools",
-	"integration-delivery",
-	"observability",
-	"networking",
-	"security",
-	"storage",
-	"streaming-messaging",
-	"other",
-}
+const MetadataFilename = "extension.yaml"
 
 type Metadata struct {
 	APIVersion string `json:"apiVersion" validate:"required"`
@@ -59,6 +45,47 @@ type Metadata struct {
 	Images               []string                                             `json:"images,omitempty"`
 	ExternalDependencies []corev1alpha1.ExternalDependency                    `json:"externalDependencies,omitempty"`
 	Annotations          map[string]string                                    `json:"annotations,omitempty"`
+}
+
+type Options struct {
+	encodeIcon bool
+}
+
+func WithEncodeIcon(encodeIcon bool) func(opts *Options) {
+	return func(opts *Options) {
+		opts.encodeIcon = encodeIcon
+	}
+}
+
+func LoadMetadata(path string, options ...func(*Options)) (*Metadata, error) {
+	opts := &Options{
+		encodeIcon: true,
+	}
+	for _, f := range options {
+		f(opts)
+	}
+
+	content, err := os.ReadFile(ospath.Join(path, MetadataFilename))
+	if err != nil {
+		return nil, err
+	}
+	metadata, err := ParseMetadata(content)
+	if err != nil {
+		return nil, err
+	}
+
+	if IsLocalFile(metadata.Icon) && opts.encodeIcon {
+		base64EncodedIcon, err := encodeIcon(ospath.Join(path, metadata.Icon))
+		if err != nil {
+			return nil, err
+		}
+		metadata.Icon = base64EncodedIcon
+	}
+
+	if err = metadata.Validate(); err != nil {
+		return nil, err
+	}
+	return metadata, nil
 }
 
 func ParseMetadata(data []byte) (*Metadata, error) {
@@ -107,7 +134,7 @@ func (md *Metadata) Validate() error {
 	return md.validateLanguageCode()
 }
 
-func (md *Metadata) ToChartYaml() (*chart.Metadata, error) {
+func (md *Metadata) ToChartYaml() *chart.Metadata {
 	var c = chart.Metadata{
 		APIVersion:   chart.APIVersionV2,
 		Name:         md.Name,
@@ -122,7 +149,14 @@ func (md *Metadata) ToChartYaml() (*chart.Metadata, error) {
 		Maintainers:  md.Maintainers,
 		Annotations:  md.Annotations,
 	}
-	return &c, nil
+	return &c
+}
+
+func DeepCopy(md *chart.Metadata) *chart.Metadata {
+	data, _ := json.Marshal(md)
+	out := &chart.Metadata{}
+	_ = json.Unmarshal(data, out)
+	return out
 }
 
 func IsLocalFile(path string) bool {
@@ -141,7 +175,7 @@ func encodeIcon(iconPath string) (string, error) {
 	}
 	var base64Encoding string
 
-	mimeType := mime.TypeByExtension(path.Ext(iconPath))
+	mimeType := mime.TypeByExtension(ospath.Ext(iconPath))
 	if mimeType == "" {
 		mimeType = http.DetectContentType(content)
 	}
@@ -149,123 +183,4 @@ func encodeIcon(iconPath string) (string, error) {
 	base64Encoding += "data:" + mimeType + ";base64,"
 	base64Encoding += base64.StdEncoding.EncodeToString(content)
 	return base64Encoding, nil
-}
-
-type Extension struct {
-	Metadata *Metadata
-	// ChartURL valid when the chart source online.
-	ChartURL string
-	// ChartData valid when the chart source local.
-	ChartData []byte
-}
-
-type ApplicationClass struct {
-	ApplicationClassGroup string               `json:"applicationClassGroup,omitempty"`
-	Name                  string               `json:"name,omitempty"`
-	Provisioner           string               `json:"provisioner,omitempty"`
-	Parameters            map[string]string    `json:"parameters,omitempty"`
-	AppVersion            string               `json:"appVersion,omitempty"`
-	PackageVersion        string               `json:"packageVersion,omitempty"`
-	Icon                  string               `json:"icon,omitempty"`
-	Description           corev1alpha1.Locales `json:"description,omitempty"`
-	Maintainer            *chart.Maintainer    `json:"maintainer,omitempty"`
-}
-
-const (
-	kubeSphereSystem  = "kubesphere-system"
-	configMapDataKey  = "chart.tgz"
-	kubeSphereManaged = "kubesphere.io/managed"
-)
-
-func (ext *Extension) ToKubernetesResources() []runtimeclient.Object {
-	var resources = []runtimeclient.Object{
-		&corev1alpha1.Extension{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: "kubesphere.io/v1alpha1",
-				Kind:       "Extension",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name: ext.Metadata.Name,
-				Labels: map[string]string{
-					corev1alpha1.CategoryLabel: ext.Metadata.Category,
-					kubeSphereManaged:          "true",
-				},
-			},
-			Spec: corev1alpha1.ExtensionSpec{
-				ExtensionInfo: corev1alpha1.ExtensionInfo{
-					Description: ext.Metadata.Description,
-					DisplayName: ext.Metadata.DisplayName,
-					Icon:        ext.Metadata.Icon,
-					Provider:    ext.Metadata.Provider,
-					Created:     metav1.Now(),
-				},
-			},
-			Status: corev1alpha1.ExtensionStatus{
-				RecommendedVersion: ext.Metadata.Version,
-			},
-		}}
-	extensionVersion := &corev1alpha1.ExtensionVersion{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "kubesphere.io/v1alpha1",
-			Kind:       "ExtensionVersion",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: fmt.Sprintf("%s-%s", ext.Metadata.Name, ext.Metadata.Version),
-			Labels: map[string]string{
-				corev1alpha1.ExtensionReferenceLabel: ext.Metadata.Name,
-				corev1alpha1.CategoryLabel:           ext.Metadata.Category,
-			},
-			Annotations: ext.Metadata.Annotations,
-		},
-		Spec: corev1alpha1.ExtensionVersionSpec{
-			InstallationMode: ext.Metadata.InstallationMode,
-			ExtensionInfo: corev1alpha1.ExtensionInfo{
-				Description: ext.Metadata.Description,
-				DisplayName: ext.Metadata.DisplayName,
-				Icon:        ext.Metadata.Icon,
-				Provider:    ext.Metadata.Provider,
-				Created:     metav1.Now(),
-			},
-			Docs:                 ext.Metadata.Docs,
-			Namespace:            ext.Metadata.Namespace,
-			Home:                 ext.Metadata.Home,
-			Keywords:             ext.Metadata.Keywords,
-			KSVersion:            ext.Metadata.KSVersion,
-			KubeVersion:          ext.Metadata.KubeVersion,
-			Sources:              ext.Metadata.Sources,
-			Version:              ext.Metadata.Version,
-			Category:             ext.Metadata.Category,
-			Screenshots:          ext.Metadata.Screenshots,
-			ExternalDependencies: ext.Metadata.ExternalDependencies,
-		},
-	}
-	if ext.ChartURL != "" {
-		extensionVersion.Spec.ChartURL = ext.ChartURL
-		resources = append(resources, extensionVersion)
-	} else {
-		configmap := &corev1.ConfigMap{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: "v1",
-				Kind:       "ConfigMap",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      fmt.Sprintf("extension-%s-%s-chart", ext.Metadata.Name, ext.Metadata.Version),
-				Namespace: kubeSphereSystem,
-			},
-			BinaryData: map[string][]byte{
-				configMapDataKey: ext.ChartData,
-			},
-		}
-		extensionVersion.Spec.ChartDataRef = &corev1alpha1.ConfigMapKeyRef{
-			Namespace: configmap.Namespace,
-			ConfigMapKeySelector: corev1.ConfigMapKeySelector{
-				LocalObjectReference: corev1.LocalObjectReference{
-					Name: configmap.Name,
-				},
-				Key: configMapDataKey,
-			},
-		}
-		resources = append(resources, extensionVersion, configmap)
-	}
-	return resources
 }
